Move and rename the namespace of the Generators (#2902)
* Rename and Move ConfigFormBaseCommandTest

Rename GeneratorConfigFormBaseCommandTest.php located in Command and
moved it to Generate as ConfigFormBaseCommandTest.php

* Rename and move ControllerCommandTest

Renamed GeneratorControllerCommandTest.php located in Command and moved
it to Generate as ControllerCommandTest.php

* Rename and move EntityCommandTest

Renamed GeneratorEntityCommandTest.php located in Command and moved it
to Generate as EntityCommandTest.php

* Rename and move EntityConfigCommandTest

Renamed GeneratorEntityConfigCommandTest.php located in Command and
moved it to Generate as EntityConfigCommandTest.php

* Rename and move EntityContentCommandTest

Renamed GeneratorEntityContentCommandTest.php located in Command and
moved it to Generate as EntityCommandCommandTest.php

* Rename and move FormCommandTest

Renamed GeneratorFormCommandTest.php located in Command and moved it to
Generate as FormCommandTest.php

* Rename and move ModuleCommandTest

Renamed GeneratorModuleCommandTest.php located in Command and moved it
to Generate as ModuleCommandTest.php

* Rename PermissionCommandTest

Renamed GeneratorPermissionCommandTest.php located in Command and moved
it to Generate as PermissionCommandTest.php

* Rename and move PluginBlockCommandTest

Renamed GeneratorPluginBlockCommandTest.php located in Command and
moved it to Generate as PluginBlockCommandTest.php

* Rename and move PluginConditionCommandTest

Renamed GeneratorPluginConditionCommandTest.php located in Command and
moved it to Generate as PluginConditionCommandTest.php

* Rename and move PluginFieldCommandTest

Renamed GeneratorPluginFieldCommandTest.php located in Command and
moved it to Generate as PluginFieldCommandTest.php

* Rename PluginFieldFormatter

Renamed GeneratorPluginFieldFormatterCommandTest.php located in Command
and moved it to Generate as PluginFieldFormatterCommandTest.php

* Rename and move PluginFieldTypeCommandTest

Renamed GeneratorPluginFieldTypeCommandTest.php located in Command and
moved it to Generate as PluginFieldTypeCommandTest.php

* Rename and move PluginFieldWidget

Renamed GeneratorPluginFieldWidgetCommandTest.php located in Command
and moved it to Generate as ControllerPluginFieldWidgetTest.php

* Rename and move PluginImageEffectCommandTest

Renamed GeneratorPluginImageEffectCommandTest.php located in Command
and moved it to Generate as PluginImageEffectCommandTest.php

* Rename and move PluginImageFormatter

Renamed GeneratorPluginImageFormatterCommandTest.php located in Command
and moved it to Generate as PluginImageFormatterCommandTest.php

* Rename and move PluginRestResourceCommandTest

Renamed GeneratorPluginRestResourceCommandTest.php located in Command
and moved it to Generate as PluginRestResourceCommandTest.php

* Rename and move PluginRulesActionCommandTest

Renamed GeneratorPluginRulesActionCommandTest.php located in Command
and moved it to Generate as PluginRulesActionCommandTest.php

* Rename and move PluginTypeAnnotationCommandTest

Renamed GeneratorPluginTypeAnnotationCommandTest.php located in Command
and moved it to Generate as PluginTypeAnnotationCommandTest.php

* Rename and move PluginTypeYamlCommandTest

Renamed GeneratorPluginTypeYamlCommandTest.php located in Command and
moved it to Generate as PluginTypeYamlCommandTest.php

* Corrected the namespace of PluginConditionCommandTest

* Rename and move ThemeCommandTest

Renamed GeneratorThemeCommandTest.php located in Command and moved it
to Generate as ThemeCommandTest.php

* Rename and move PluginTypeYaml

Renamed GeneratorPluginTypeYamlCommandTest.php located in Command and
moved it to Generate as PluginTypeYamlCommandTest.php

* Fix the class erasing Generator
