+++ Test/Command/Generate/ConfigFormBaseCommandTest.php
@@ -4,13 +4,13 @@
  * Contains \Drupal\Console\Test\Command\GeneratorConfigFormBaseCommandTest.
  */
 
-namespace Drupal\Console\Test\Command;
+namespace Drupal\Console\Test\Command\Generate;
 
 use Drupal\Console\Command\Generate\ConfigFormBaseCommand;
 use Symfony\Component\Console\Tester\CommandTester;
 use Drupal\Console\Test\DataProvider\ConfigFormBaseDataProviderTrait;
 
-class GeneratorConfigFormBaseCommandTest extends GenerateCommandTest
+class ConfigFormBaseCommandTest extends GenerateCommandTest
 {
     use ConfigFormBaseDataProviderTrait;
