+++ Test/Command/Generate/PluginConditionCommandTest.php
@@ -4,16 +4,16 @@
  * Contains \Drupal\Console\Test\Command\GeneratorPluginConditionCommandTest.
  */
 
-namespace Drupal\Console\Test\Command;
+namespace Drupal\Console\Test\Command\Generate;
 
 use Drupal\Console\Command\Generate\PluginConditionCommand;
 use Symfony\Component\Console\Tester\CommandTester;
 use Drupal\Console\Test\DataProvider\PluginConditionDataProviderTrait;
 
-class GeneratorPluginConditionCommandTest extends GenerateCommandTest
+class PluginConditionCommandTest extends GenerateCommandTest
 {
     use PluginConditionDataProviderTrait;
-    
+
     /**
      * Plugin block generator test
      *
