+++ Test/Command/Generate/PluginTypeYamlCommandTest.php
@@ -1,16 +1,16 @@
 <?php
 /**
  * @file
- * Contains \Drupal\Console\Test\Command\GeneratorPluginTypeAnnotationCommandTest.
+ * Contains \Drupal\Console\Test\Command\GeneratorPluginTypeYamlCommandTest.
  */
 
-namespace Drupal\Console\Test\Command;
+namespace Drupal\Console\Test\Command\Generate;
 
-use Drupal\Console\Command\Generate\PluginTypeAnnotationCommand;
+use Drupal\Console\Command\Generate\PluginTypeYamlCommand;
 use Symfony\Component\Console\Tester\CommandTester;
 use Drupal\Console\Test\DataProvider\PluginTypeYamlDataProviderTrait;
 
-class GeneratorPluginTypeAnnotationCommandTest extends GenerateCommandTest
+class PluginTypeYamlCommandTest extends GenerateCommandTest
 {
     use PluginTypeYamlDataProviderTrait;
 
@@ -18,19 +18,19 @@ class GeneratorPluginTypeAnnotationCommandTest extends GenerateCommandTest
      * Plugin type yaml generator test
      *
      * @param $module
-     * @param $class_name
-     * @param $machine_name
-     * @param $label
+     * @param $plugin_class
+     * @param $plugin_name
+     * @param $plugin_file_name
      *
      * @dataProvider commandData
      */
     public function testGeneratePluginTypeYaml(
         $module,
-        $class_name,
-        $machine_name,
-        $label
+        $plugin_class,
+        $plugin_name,
+        $plugin_file_name
     ) {
-        $command = new PluginTypeAnnotationCommand($this->getHelperSet());
+        $command = new PluginTypeYamlCommand($this->getHelperSet());
         $command->setHelperSet($this->getHelperSet());
         $command->setGenerator($this->getGenerator());
 
@@ -39,9 +39,9 @@ class GeneratorPluginTypeAnnotationCommandTest extends GenerateCommandTest
         $code = $commandTester->execute(
             [
               '--module'            => $module,
-              '--class'        => $class_name,
-              '--machine-name'      => $machine_name,
-              '--label'             => $label
+              '--class'        => $plugin_class,
+              '--plugin-name'       => $plugin_name,
+              '--plugin-file-name'  => $plugin_file_name
             ],
             ['interactive' => false]
         );
