+++ Test/Command/Generate/ThemeCommandTest.php
@@ -4,16 +4,16 @@
  * Contains Drupal\Console\Test\Command\GeneratorThemeCommandTest.
  */
 
-namespace Drupal\Console\Test\Command;
+namespace Drupal\Console\Test\Command\Generate;
 
 use Drupal\Console\Command\Generate\ThemeCommand;
 use Symfony\Component\Console\Tester\CommandTester;
 use Drupal\Console\Test\DataProvider\ThemeDataProviderTrait;
 
-class GeneratorThemeCommandTest extends GenerateCommandTest
+class ThemeCommandTest extends GenerateCommandTest
 {
     use ThemeDataProviderTrait;
-    
+
     /**
      * Theme generator test
      *
